booksim v2.0: add the nvidia dgx1 topology
adding the nvidia dgx1 topology and  modifing the trafficmananger to support dgx1. But, the efficiency is not ideal.

+++ booksim2/src/booksim_wrapper.hpp
@@ -52,7 +52,7 @@ namespace booksim{
 			int CheckInjectionQueue(int source, int cl);
 
 			//更新周期计数
-			void UpdateSimTime(int cycles);
+			void UpdateSimTime(long long cycles);
 
 			//打印数据，更新截止到当前的数据并打印它们
 			void PrintStats(std::ostream & os);
